adversarial fiddling

+++ Code/Models/Attention_seq2seq.py
@@ -402,7 +402,7 @@ class _Seq2Seq(nn.Module):
             # insert dec_input token embedding, previous hidden state and all encoder hidden states
             # receive output tensor (predictions) and new hidden state
             #output, hidden = self.decoder(dec_input, hidden, encoder_outputs)
-            output, hidden, a_ = self.decoder(dec_input, hidden.double(), encoder_outputs, mask)
+            output, hidden, a_ = self.decoder(dec_input, hidden, encoder_outputs, mask)
             # cleaning
             del a_
             # place predictions in a tensor holding predictions for each token
